ebarristers/requirements.txt
python 3.7
django 3.0.5
django-crispy-forms-1.9.0
django-phonenumber-field-4.0.0
phonenumbers-8.12.1
pycryptodome-3.9.7

